+++ website/modules/planets/jupiter.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Jupiter } from '@thecuvii/solaris/jupiter'
+import { Jupiter } from '@cuvii/solaris/jupiter'
 
 import { usePlanetPreviewProps } from '../planet-page/use-planet-preview-props'
 import { textures } from '../showcase/showcase-data'