+++ website/modules/planets/pluto.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Pluto } from '@thecuvii/solaris/pluto'
+import { Pluto } from '@cuvii/solaris/pluto'
 
 import { usePlanetPreviewProps } from '../planet-page/use-planet-preview-props'
 import { textures } from '../showcase/showcase-data'