+++ website/modules/planets/sun.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Sun } from '@thecuvii/solaris/sun'
+import { Sun } from '@cuvii/solaris/sun'
 
 import { usePlanetPreviewProps } from '../planet-page/use-planet-preview-props'
 import { textures } from '../showcase/showcase-data'