+++ website/modules/planets/uranus.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Uranus } from '@thecuvii/solaris/uranus'
+import { Uranus } from '@cuvii/solaris/uranus'
 
 import { usePlanetPreviewProps } from '../planet-page/use-planet-preview-props'